Extracted where to buy button out
Cleaned up some of the spaghetti from the where to buy button.  This is only working on mobile phones and breaks at an early breakpoint.  Need to make sure the "click to follow" hover button even makes sense to have for mobile.  For the time being I'm going to leave the script tags inline since I'm not 100% sure how the asset pipeline deals with JS.  Once this chunk is done mobile page should be good to go.